+++ reptor/models/Base.py
@@ -106,6 +106,12 @@ class BaseModel:
                     pass
         return dict_values
 
+    def __repr__(self):
+        return f"{self.__class__.__name__}(id={self.id})"
+    
+    def __str__(self):
+        return self.id
+
 
 class ProjectFieldTypes(enum.Enum):
     cvss = "cvss"