+++ reptor/models/Note.py
@@ -8,8 +8,13 @@ class NoteBase(BaseModel):
     text: str = ""
     checked: typing.Optional[bool] = None
     icon_emoji: str = ""
-    status_emoji: str = ""
     order: int = 0
+
+    def __str__(self):
+        return self.title
+    
+    def __repr__(self):
+        return f'{self.__class__.__name__}(title="{self.title}", id="{self.id}")'
 
 
 class Note(NoteBase):
@@ -22,18 +27,20 @@ class Note(NoteBase):
         note.text = note_template.text
         note.checked = note_template.checked
         note.icon_emoji = note_template.icon_emoji
-        note.status_emoji = note_template.status_emoji
         note.order = note_template.order
         if note_template.parent:
             note.parent = note_template.parent
         return note
+    
+    def __repr__(self):
+        return f'Note(title="{self.title}", id="{self.id}", parent="{self.parent}")'
 
 
 class NoteTemplate(NoteBase):
     template: str = ""
     template_data: dict = dict()
-    parent: typing.Optional[str] = None  # Parent ID preferred over parent_notetitle
-    parent_notetitle: typing.Optional[str] = None
+    parent: typing.Optional[str] = None  # parent preferred over parent_title
+    parent_title: typing.Optional[str] = None
     children: typing.List["NoteTemplate"] = list()
     force_new: bool = False  # Create a new note even if same title exists
 
@@ -45,3 +52,6 @@ class NoteTemplate(NoteBase):
     @classmethod
     def from_kwargs(cls, **kwargs):
         return cls(data=kwargs)
+    
+    def __repr__(self):
+        return f'NoteTemplate(title="{self.title}", id="{self.id}", parent="{self.parent}")'