+++ reptor/models/User.py
@@ -34,6 +34,7 @@ class User(BaseModel):
 
     username: str = ""
     name: str = ""
+    color: str = ""
     title_before: str = ""
     first_name: str = ""
     middle_name: str = ""
@@ -55,3 +56,9 @@ class User(BaseModel):
     is_mfa_enabled: bool = False
     can_login_local: bool = False
     can_login_sso: bool = False
+
+    def __str__(self):
+        return f"{self.name} ({self.username})"
+
+    def __repr__(self):
+        return f'User(username="{self.username}", name="{self.name}", email="{self.email}", id="{self.id}")'