+++ reptor/plugins/projects/Project/tests/test_integration_project.py
@@ -53,8 +53,7 @@ class TestIntegrationProject(object):
         )
         p.communicate()
         assert p.returncode == 0
-        del projects_api._project_dict
-        assert projects_api._get_project().readonly is True
+        assert projects_api.fetch_project().readonly is True
 
         # Reactivate project
         p = subprocess.Popen(
@@ -62,8 +61,7 @@ class TestIntegrationProject(object):
         )
         p.communicate()
         assert p.returncode == 0
-        del projects_api._project_dict
-        assert projects_api._get_project().readonly is False
+        assert projects_api.fetch_project().readonly is False
 
     def test_export_tar_gz(self):
         fname = "myproject.tar.gz"