+++ reptor/plugins/projects/PushProject/PushProject.py
@@ -4,12 +4,13 @@ import json
 import sys
 import typing
 
+import requests
 import tomli
 
 from reptor.lib.plugins.UploadBase import UploadBase
-from reptor.models.Section import Section as SectionModel
 from reptor.models.Finding import Finding as FindingModel
 from reptor.models.ProjectDesign import ProjectDesign
+from reptor.models.Section import Section as SectionModel
 
 
 class PushProject(UploadBase):
@@ -37,6 +38,7 @@ class PushProject(UploadBase):
         parser.add_argument("projectdata", nargs="?", type=argparse.FileType("rb"))
 
     def run(self):
+        ##### DEPRECATED
         len_sections = len(
             self.reptor.api.projects.update_report_fields(
                 self.projectdata.get("report_data", {})
@@ -44,27 +46,52 @@ class PushProject(UploadBase):
         )
         if len_sections:
             self.log.success(
-                f"Updated {len_sections} report section{'s'[:len_sections^1]}."
+                f"Updated {len_sections} report section{'s'[:len_sections^1]} using DEPRECATED \"report_data\"."
+            )
+        ##### END DEPRECATED
+
+        len_sections = len(
+            self.reptor.api.projects.update_sections(
+                sections=self.projectdata.get("sections", [])
+            )
+        )
+        if len_sections:
+            self.log.success(
+                f"Updated {len_sections} section{'s'[:len_sections^1]}."
             )
         else:
-            self.log.display("No report sections updated.")
+            self.log.display("No sections updated.")
 
         # Check for valid finding field data format
         project_design = self.reptor.api.project_designs.project_design
         findings = list()
-        for finding in self.projectdata.get("findings", []):  # Check data format
+        for finding in self.projectdata.get("findings", []):
             findings.append(
                 (
                     finding,
                     FindingModel(
                         finding, project_design, strict_type_check=False
-                    ),
+                    ),  # Checks data format
                 )
             )
         # Upload
         for finding, model in findings:
+            if model.id:
+                # Update existing finding
+                try:
+                    self.reptor.api.projects.update_finding(model.id, finding)
+                    self.log.success(f'Updated finding "{model.data.title.value}".')
+                except requests.exceptions.HTTPError as e:
+                    if e.response.status_code == 404:
+                        # Finding with finding id not found
+                        # We'll try to create the finding
+                        pass
+                    else:
+                        raise e
+                else:
+                    continue
             self.reptor.api.projects.create_finding(finding)
-            self.log.success(f'Created finding "{model.data.title}".')
+            self.log.success(f'Created finding "{model.data.title.value}".')
         if not findings:
             self.log.display("No findings created.")
 
@@ -85,11 +112,23 @@ class PushProject(UploadBase):
 
         # Create model to assure compatibility of predefined fields
         assert isinstance(loaded_content, dict)
+
+        ##### DEPRECATED
+        if "report_data" in loaded_content:
+            self.log.warning("The 'report_data' field is deprecated and will be removed in a future version. Use 'sections' instead.")
         report_data = loaded_content.get("report_data", {})
         assert isinstance(report_data, dict)
         SectionModel(
             {"data": report_data}, ProjectDesign(), strict_type_check=False
         )
+        ##### END DEPRECATED
+
+        sections = loaded_content.get("sections", [])
+        assert isinstance(sections, list)
+        for section in sections:
+            assert isinstance(section, dict)
+            SectionModel(section, ProjectDesign(), strict_type_check=False)
+
         findings = loaded_content.get("findings", [])
         assert isinstance(findings, list)
         for finding in findings: