+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-# Upcoming
+# 0.32
 * Fix bug that removes "insecure" option from conf file
-* Allow updating findings
+* Allow updating findings with "findings" plugin
+* Allow updating findings with "pushproject" plugin
+* Deprecate "report_data" in favor of "sections" in "pushproject" plugin
+* Changes in APIs for Python library usage
 * Clean up NOTICE file
 
 # 0.31